Update packages, reorganize SwipeCardView, and refactor code
Updated `Microsoft.Maui.Controls` to version `8.0.92` and
`Microsoft.Extensions.Logging.Debug` to version `8.0.1` in
project files. Added `Microsoft.Maui.Controls.Compatibility`
package reference. Reorganized `SwipeCardView.cs` with regions
for better code structure. Moved `OnPanUpdated` to `Event
Handlers` region and added a new method for pan gesture updates.
Removed `DraggingCardPosition` property, initializing it in the
constructor. Moved `SwipeCardDirectionExtensions` to a separate
file and cleaned up redundant code and `using` directives.
